Mainクラス内のシューティング部分をStageクラスに移植
